separate API operations into modules
